M3-112 Refactor : widgets 폴더 분리
